refactor: rename modules/ to handlers/ in backend
