+++ apps/backend/src/handlers/auth/currentUser.ts
