+++ apps/backend/src/handlers/auth/logout.ts
