+++ apps/backend/src/handlers/catalog/song/delete.ts
